requirements.txt
alembic==0.9.6
appdirs==1.4.3
argcomplete==1.9.3
arrow==0.12.0
bcrypt==3.1.4
binaryornot==0.4.4
blinker==1.4
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
cookiecutter==1.6.0
Flask==0.12.2
Flask-Admin==1.5.0
Flask-Bcrypt==0.7.1
Flask-Caching==1.3.3
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-Webpack==0.1.0
Flask-WTF==0.14.2
future==0.16.0
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jinja2-time==0.2.0
Mako==1.0.7
MarkupSafe==1.0
mysqlclient==1.3.12
npm==0.1.1
optional-django==0.1.0
Pillow==5.0.0
poyo==0.4.1
psycopg2==2.7.3.2
pycparser==2.18
pyserial==3.4
python-dateutil==2.6.1
python-editor==1.0.3
python-escpos==2.2.0
pyusb==1.0.2
PyYAML==3.12
qrcode==5.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.0
urllib3==1.22
Werkzeug==0.13
whichcraft==0.4.1
WTForms==2.1
